Revert "Menambahkan model Job"
This reverts commit bd889a7ce82c9531763bbce917e10709204bfdf1.
