--- .gitignore
@@ -1,2 +1,0 @@
-/MBKM_JDBC/MBKM_JDBC/nbproject/private/
-/MBKM_JDBC/MBKM_JDBC/build/
