+++ MBKM_JDBC/manifest.mf
