+++ MBKM_JDBC/nbproject/private/private.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/C:/Users/Asus/OneDrive/Documents/NetBeansProjects/MBKM_JDBC/src/daos/RegionDAO.java</file>
+            <file>file:/C:/Users/Asus/OneDrive/Documents/NetBeansProjects/MBKM_JDBC/src/tools/MBKM_JDBC.java</file>
+        </group>
+    </open-files>
+</project-private>
